ISY994 Structure updates in prep for PyISYv2 (Part 2)
- No breaking changes in this PR.
- Move core Entity Class definitions to separate file.
- New ISYNodeEntity and ISYProgramEntity base classes.
- Update inheriances on other classes.
- Move helper functions to separate file.
